requirements.txt
# requirements.txt
streamlit
azure-identity
azure-ai-projects
dotenv
azure-core
azure-ai-agents